separated files and removed deps
