+++ packages/define-remix-app/package.json
@@ -13,9 +13,6 @@
     "@remix-run/testing": "^2.11.2",
     "@wixc3/app-core": "^3.0.0"
   },
-  "devDependencies": {
-    "@wixc3/app-core-test-kit": "^3.0.0"
-  },
   "files": [
     "dist",
     "!dist/test",
